requirements.txt
Flask
Pillow
Flask-SocketIO
mysql-connector-python